requirements.txt
uvicorn~=0.18.3
fastapi~=0.84.0
pydantic~=1.10.2
SQLAlchemy~=1.4.41
starlette~=0.19.1
passlib~=1.7.4
python-jose